+++ R/mod_scatter.R
@@ -86,13 +86,16 @@ SP <- poc( # nolint
 #' It also includes a set of listings with information about the population and the regression and correlation
 #' estimates.
 #'
+#' @inheritParams scatterplot_server
+#'
 #' @name mod_scatterplot
 #'
 #' @keywords main
 #'
 NULL
 
-#' @describeIn mod_scatterplot UI
+#' Scatter plot UI function
+#' @keywords developers
 #' @param id Shiny ID `[character(1)]`
 #' @export
 scatterplot_UI <- function(id) { # nolint
@@ -194,7 +197,9 @@ scatterplot_UI <- function(id) { # nolint
   }
 }
 
-#' @describeIn mod_scatterplot Server
+#' Scatter plot server function
+#'
+#' @keywords developers
 #'
 #' @description
 #'
@@ -259,7 +264,7 @@ scatterplot_server <- function(id,
                                dataset_name = shiny::reactive(character(0)),
                                cat_var = "PARCAT",
                                par_var = "PARAM",
-                               value_vars = c("AVAL", "PCHG"),
+                               value_vars = "AVAL",
                                visit_var = "AVISIT",
                                subjid_var = "SUBJID",
                                default_x_cat = NULL,
@@ -702,7 +707,7 @@ mod_scatterplot <- function(module_id,
                             group_dataset_name,
                             cat_var = "PARCAT",
                             par_var = "PARAM",
-                            value_vars = c("AVAL", "PCHG"),
+                            value_vars = "AVAL",
                             visit_var = "AVISIT",
                             subjid_var = "SUBJID",
                             default_x_cat = NULL,